Added consumes calls to StripCPEESProducer
-Switched to using ESGetTokens
-Removed 3 get calls to data which were never used
-Consolidated header and source into one file
