+++ RecoLocalTracker/SiStripRecHitConverter/plugins/StripCPEESProducer.cc
@@ -1,4 +1,12 @@
-#include "RecoLocalTracker/SiStripRecHitConverter/plugins/StripCPEESProducer.h"
+#include "FWCore/Framework/interface/ESProducer.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "FWCore/Framework/interface/ModuleFactory.h"
+#include "FWCore/Framework/interface/ESHandle.h"
+#include "RecoLocalTracker/Records/interface/TkStripCPERecord.h"
+#include "RecoLocalTracker/ClusterParameterEstimator/interface/StripClusterParameterEstimator.h"
+
+#include "CalibTracker/Records/interface/SiStripDependentRecords.h"
+
 #include "RecoLocalTracker/SiStripRecHitConverter/interface/StripCPE.h"
 #include "RecoLocalTracker/SiStripRecHitConverter/interface/StripCPEfromTrackAngle.h"
 #include "RecoLocalTracker/SiStripRecHitConverter/interface/StripCPEgeometric.h"
@@ -7,10 +15,25 @@
 #include "CondFormats/SiStripObjects/interface/SiStripBackPlaneCorrection.h"
 #include "CondFormats/SiStripObjects/interface/SiStripConfObject.h"
 #include "CondFormats/SiStripObjects/interface/SiStripLatency.h"
-#include "CondFormats/SiStripObjects/interface/SiStripNoises.h"
-#include "CondFormats/SiStripObjects/interface/SiStripApvGain.h"
-#include "CondFormats/SiStripObjects/interface/SiStripBadStrip.h"
-#include "FWCore/Framework/interface/ESHandle.h"
+
+class StripCPEESProducer : public edm::ESProducer {
+public:
+  StripCPEESProducer(const edm::ParameterSet&);
+  std::unique_ptr<StripClusterParameterEstimator> produce(const TkStripCPERecord&);
+
+private:
+  enum CPE_t { SIMPLE, TRACKANGLE, GEOMETRIC, FAKE };
+
+  CPE_t cpeNum;
+  edm::ParameterSet parametersPSet;
+
+  edm::ESGetToken<TrackerGeometry, TrackerDigiGeometryRecord> pDDToken_;
+  edm::ESGetToken<MagneticField, IdealMagneticFieldRecord> magfieldToken_;
+  edm::ESGetToken<SiStripLorentzAngle, SiStripLorentzAngleDepRcd> lorentzAngleToken_;
+  edm::ESGetToken<SiStripBackPlaneCorrection, SiStripBackPlaneCorrectionDepRcd> backPlaneCorrectionToken_;
+  edm::ESGetToken<SiStripConfObject, SiStripConfObjectRcd> confObjToken_;
+  edm::ESGetToken<SiStripLatency, SiStripLatencyRcd> latencyToken_;
+};
 
 StripCPEESProducer::StripCPEESProducer(const edm::ParameterSet& p) {
   std::string name = p.getParameter<std::string>("ComponentName");
@@ -22,6 +45,7 @@ StripCPEESProducer::StripCPEESProducer(const edm::ParameterSet& p) {
   else
     type = p.getParameter<std::string>("ComponentType");
 
+  std::map<std::string, CPE_t> enumMap;
   enumMap[std::string("SimpleStripCPE")] = SIMPLE;
   enumMap[std::string("StripCPEfromTrackAngle")] = TRACKANGLE;
   enumMap[std::string("StripCPEgeometric")] = GEOMETRIC;
@@ -31,45 +55,39 @@ StripCPEESProducer::StripCPEESProducer(const edm::ParameterSet& p) {
 
   cpeNum = enumMap[type];
   parametersPSet = (p.exists("parameters") ? p.getParameter<edm::ParameterSet>("parameters") : p);
-  setWhatProduced(this, name);
+  setWhatProduced(this, name)
+      .setConsumes(pDDToken_)
+      .setConsumes(magfieldToken_)
+      .setConsumes(lorentzAngleToken_)
+      .setConsumes(backPlaneCorrectionToken_)
+      .setConsumes(confObjToken_)
+      .setConsumes(latencyToken_);
 }
 
 std::unique_ptr<StripClusterParameterEstimator> StripCPEESProducer::produce(const TkStripCPERecord& iRecord) {
-  edm::ESHandle<TrackerGeometry> pDD;
-  iRecord.getRecord<TrackerDigiGeometryRecord>().get(pDD);
-  edm::ESHandle<MagneticField> magfield;
-  iRecord.getRecord<IdealMagneticFieldRecord>().get(magfield);
-  edm::ESHandle<SiStripLorentzAngle> lorentzAngle;
-  iRecord.getRecord<SiStripLorentzAngleDepRcd>().get(lorentzAngle);
-  edm::ESHandle<SiStripBackPlaneCorrection> backPlaneCorrection;
-  iRecord.getRecord<SiStripBackPlaneCorrectionDepRcd>().get(backPlaneCorrection);
-  edm::ESHandle<SiStripConfObject> confObj;
-  iRecord.getRecord<SiStripConfObjectRcd>().get(confObj);
-  edm::ESHandle<SiStripLatency> latency;
-  iRecord.getRecord<SiStripLatencyRcd>().get(latency);
-  edm::ESHandle<SiStripNoises> noise;
-  iRecord.getRecord<SiStripNoisesRcd>().get(noise);
-  edm::ESHandle<SiStripApvGain> gain;
-  iRecord.getRecord<SiStripApvGainRcd>().get(gain);
-  edm::ESHandle<SiStripBadStrip> bad;
-  iRecord.getRecord<SiStripBadChannelRcd>().get(bad);
+  TrackerGeometry const& pDD = iRecord.get(pDDToken_);
+  MagneticField const& magfield = iRecord.get(magfieldToken_);
+  SiStripLorentzAngle const& lorentzAngle = iRecord.get(lorentzAngleToken_);
+  SiStripBackPlaneCorrection const& backPlaneCorrection = iRecord.get(backPlaneCorrectionToken_);
+  SiStripConfObject const& confObj = iRecord.get(confObjToken_);
+  SiStripLatency const& latency = iRecord.get(latencyToken_);
 
   std::unique_ptr<StripClusterParameterEstimator> cpe;
 
   switch (cpeNum) {
     case SIMPLE:
       cpe = std::make_unique<StripCPE>(
-          parametersPSet, *magfield, *pDD, *lorentzAngle, *backPlaneCorrection, *confObj, *latency);
+          parametersPSet, magfield, pDD, lorentzAngle, backPlaneCorrection, confObj, latency);
       break;
 
     case TRACKANGLE:
       cpe = std::make_unique<StripCPEfromTrackAngle>(
-          parametersPSet, *magfield, *pDD, *lorentzAngle, *backPlaneCorrection, *confObj, *latency);
+          parametersPSet, magfield, pDD, lorentzAngle, backPlaneCorrection, confObj, latency);
       break;
 
     case GEOMETRIC:
       cpe = std::make_unique<StripCPEgeometric>(
-          parametersPSet, *magfield, *pDD, *lorentzAngle, *backPlaneCorrection, *confObj, *latency);
+          parametersPSet, magfield, pDD, lorentzAngle, backPlaneCorrection, confObj, latency);
       break;
 
     case FAKE:
@@ -79,3 +97,5 @@ std::unique_ptr<StripClusterParameterEstimator> StripCPEESProducer::produce(cons
 
   return cpe;
 }
+
+DEFINE_FWK_EVENTSETUP_MODULE(StripCPEESProducer);
